+++ tool/istarcore/istarcore.js
@@ -459,22 +459,22 @@ var istar = function () {
             "use strict";
             //TODO prevent repeated links
 
-            var link;
+            var link1;
             if (depender.isKindOfActor()) {
-                link = new joint.shapes.istar.DependencyLink({
+                link1 = new joint.shapes.istar.DependencyLink({
                     'source': {id: depender.id, selector: 'circle'},
                     'target': {id: dependum.id}
                 });
             }
             else {
-                link = new joint.shapes.istar.DependencyLink({
+                link1 = new joint.shapes.istar.DependencyLink({
                     'source': {id: depender.id},
                     'target': {id: dependum.id}
                 });
             }
 
-            istar.graph.addCell(link);
-            _updateLinkLabelRotation(link, depender, dependum);
+            istar.graph.addCell(link1);
+            _updateLinkLabelRotation(link1, depender, dependum);
 
             var link2;
             if (dependee.isKindOfActor()) {
@@ -492,6 +492,9 @@ var istar = function () {
             istar.graph.addCell(link2);
             _updateLinkLabelRotation(link2, dependum, dependee);
 
+            link1.prop('otherHalf', link2);
+            link2.prop('otherHalf', link1);
+
             dependum.prop('isDependum', true);
             var dependumPosition = {x: 50, y: 50};
             dependumPosition.x = (depender.prop('position/x') + dependee.prop('position/x')) / 2;
@@ -499,7 +502,7 @@ var istar = function () {
             dependum.prop('position', dependumPosition);
 
             //move links to the back, so that they don't appear on top of the element's shape
-            link.toBack();
+            link1.toBack();
             link2.toBack();
             //move all the actors even further back, so that they don't impede the visualization of the dependency links
             var actors = _.filter(istar.graph.getElements(), function (element) {
@@ -508,7 +511,7 @@ var istar = function () {
             _.each(actors, function (actor) {
                 actor.toBack();
             });
-            return [link, link2];
+            return [link1, link2];
         },
         addLinkBetweenNodes: function (linkName, shape, source, target, value) {
             "use strict";