+++ CHANGELOG.md
@@ -18,7 +18,8 @@ Documentation added accordingly (docs/EVENTS.md.
 view got wider and covered the name.
 This bug is purely aesthetical.
 The view got restored after any (de)selection event.
-This is just a workaround, a proper fix is still needed. 
+This is just a workaround, a proper fix is still needed.
+- Now it deletes the entire dependency link when deleting a depender or a dependee, preventing dangling dependums
 
 ## [Release v1.0.1] - 2018-03-05
 ### Added